refactor: move utility functions
