+++ e2e-app/e2e/getElementText.ts
@@ -1,0 +1,6 @@
+import { IndexableNativeElement } from 'detox/detox'
+
+export async function getElementText(element: IndexableNativeElement): Promise<string> {
+  const attributes = (await element.getAttributes()) as Record<string, any>
+  return attributes?.text ?? attributes?.label ?? ''
+}
