+++ e2e-app/e2e/wait.ts
@@ -1,0 +1,3 @@
+export function wait(ms: number) {
+  return new Promise((resolve) => setTimeout(resolve, ms))
+}
